Fix learn-rate schedule:
Set up a new hyperparameters (eta) to scale a fixed learning rate
hyperparameter. Fixes issue with weight decay falling along with the
learning rate.

Also, have learn_rate scale the gradient only and add a new weight decay
hyperparameter for scaling in AdamW.

Change the names of a bunch of hyperparameters to be more consistent.